chore: move clickhouse config into app config for consistency
