--- apps/backend/src/clickhouse/clickhouse.config.ts
@@ -1,18 +1,0 @@
-export interface IClickhouseConfig {
-  /**
-   * ClickHouse connection URL. Must include the database in the path.
-   * Example: http://default:password@host:8123/dealbot
-   * If unset, ClickHouse emission is disabled.
-   */
-  url: string | undefined;
-  batchSize: number;
-  flushIntervalMs: number;
-}
-
-export function loadClickhouseConfig(): IClickhouseConfig {
-  return {
-    url: process.env.CLICKHOUSE_URL || undefined,
-    batchSize: Number.parseInt(process.env.CLICKHOUSE_BATCH_SIZE || "500", 10),
-    flushIntervalMs: Number.parseInt(process.env.CLICKHOUSE_FLUSH_INTERVAL_MS || "5000", 10),
-  };
-}
